feat: responsive skills and achievements section design
This is an inital commit outline the mockup and basic html markup to create
responsive design of achievements and certification section, and skills section.
Most of the markup is semantic but this commit is not the final design. Further
commits can be done on skills and achievements section design to improve the
markup and design of this section!

This section also need to implement the necessary javascript to make the progressbar
animation working.

Furthermore, JQuery library is included to fix the error caused by bootstrap.js
(Error: JQuery is required in bootstrap.js)